documentation: (Toy) use FileLineColLoc in Toy
